updated frequent syntax
