+++ ProjectTwo/.bsp/sbt.json
@@ -1,1 +1,1 @@
-{"name":"sbt","version":"1.6.0","bspVersion":"2.0.0-M5","languages":["scala"],"argv":["C:\\Program Files\\Eclipse Adoptium\\jdk-8.0.312.7-hotspot\\jre/bin/java","-Xms100m","-Xmx100m","-classpath","C:/Users/emerg/AppData/Roaming/JetBrains/IdeaIC2021.3/plugins/Scala/launcher/sbt-launch.jar","xsbt.boot.Boot","-bsp","--sbt-launch-jar=C:/Users/emerg/AppData/Roaming/JetBrains/IdeaIC2021.3/plugins/Scala/launcher/sbt-launch.jar"]}
+{"name":"sbt","version":"1.6.0","bspVersion":"2.0.0-M5","languages":["scala"],"argv":["C:\\Program Files\\Java\\jdk-17.0.1/bin/java","-Xms100m","-Xmx100m","-classpath","C:/Users/brady/AppData/Roaming/JetBrains/IdeaIC2021.2/plugins/Scala/launcher/sbt-launch.jar","xsbt.boot.Boot","-bsp","--sbt-launch-jar=C:/Users/brady/AppData/Roaming/JetBrains/IdeaIC2021.2/plugins/Scala/launcher/sbt-launch.jar"]}
