+++ ProjectTwo/src/main/scala/DataPractice - Copy.scala
@@ -1,0 +1,52 @@
+import io.netty.util.ResourceLeakDetector.Level
+import org.apache.spark.sql._
+import org.apache.spark.sql.hive.test.TestHive.sparkContext
+import org.apache.spark.sql.functions._
+
+import sys.process._
+
+object DataPractice {
+
+  def main(args: Array[String]): Unit = {
+
+
+    System.setProperty("hadoop.home.dir", "C:\\Hadoop")
+    val spark = SparkSession
+      .builder
+      .appName("Hello Hive")
+      .config("spark.master", "local")
+      .enableHiveSupport()
+      .getOrCreate()
+    println("created spark session")
+
+    spark.sparkContext.setLogLevel("ERROR")
+
+    spark.sql("DROP TABLE IF EXISTS TotalConfirmed")
+    spark.sql("CREATE TABLE TotalConfirmed(" +
+      "SNo INT," +
+      "ObservationDate STRING," +
+      "ProvinceorState STRING," +
+      "CountryorRegion STRING," +
+      "LastUpdate STRING," +
+      "Confirmed DOUBLE," +
+      "Deaths DOUBLE," +
+      "Recovered DOUBLE)" +
+      "row format delimited fields terminated by ','" +
+      "stored as textfile"
+    )
+
+    spark.sql(
+      "LOAD DATA LOCAL INPATH 'covid_19_data_complete(Kaggle).csv'" +
+      "OVERWRITE INTO TABLE TotalConfirmed"
+    )
+
+    spark.sql("SELECT FLOOR(sum(Confirmed)) as TotalConfirmed FROM TotalConfirmed")
+      .show()
+
+    spark.sql("SELECT CountryorRegion, " +
+      "MAX(Deaths) / MAX(Confirmed) AS country_mortality_rate " +
+      "FROM TotalConfirmed WHERE CountryorRegion = 'Peru' OR CountryorRegion = 'Bulgaria' OR CountryorRegion GROUP BY CountryorRegion ORDER BY 2 DESC").show()
+
+    spark.close()
+  }
+}
